requirements.txt
defusedxml
python-dotenv
